requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==0.19.1

profanityfilter>=2.0.6